fix: added transaction type
